refactor: :truck: rename to `extensions.py` from `custom_checks.py`
